💅 Update page components to use formatting utilities
